pyproject.toml
[project]
name = "astra-fin"
version = "0.2.0"
description = "The ONE Financial AI Agent"
requires-python = ">=3.11"
dependencies = [

    "langgraph",
    "langchain",
    "langchain-ollama",
    "langchain-community",
    "pydantic",
    "python-dotenv",
    "qdrant-client",
    "langchain-qdrant",
    "chainlit",
    "httpx",
    "yfinance"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black"
]

[tool.setuptools]
packages = ["src"]